+++ NEWS.md
@@ -1,8 +1,26 @@
 # sooty dev
 
-* Renamed package
+* dataset() is now deprecated in favour of datasource()
+ 
+* Created standalone function available_datasets(), and removed `@curated`, `@available_datasets` from the datasource object. 
+
+* Now using more robust model of GDAL protocol and absolute URLs. 
+
+* Added an S7 object to set dataset id and retrieve files. 
+
+* Removed all read functionality. 
+
+* Removed memoization. 
+
+* Fix scaling mistake for AMSR ice. 
 
 * Added 3.125km AMSR ice. 
+
+# sooty 0.1.0 
+
+* Initial CRAN version. 
+
+* Renamed package. 
 
 # idt 0.0.1